Added a bin/energyfaker:  a script to imitate the behavior of a Rhizome device, using a device "profile" as generated by dev/sql_to_fake_rhizome_profiles.py.  This is useful for testing.
Added a *very* rough 24hr-graph to the web display.  It works well
enough, but would a graph for a given day (as opposed to the last 24
hours) be more useful or desirable?  It might be good to make e.g. a
per-two-minute average for 24-hour display...

Made other minor modifications.  For instance, the port for each
sensor is now stored in the database (so that ports can now differ by
sensor).